Stop calling it HeaderPrototype since it's not a prototype anymore
